split out ToolbarPluginFormControl
